configLoader rename
