+++ index.js
@@ -1,4 +1,4 @@
-require('./modules/loadConfig')
+require('./modules/configLoader')
 
 const server = require('./modules/server')
 const publicServer = require('./modules/publicServer')
